+++ lib/features/point_of_sale/presentation/screens/screns.dart
@@ -1,1 +1,2 @@
 export 'point_of_sale_screen.dart';
+export 'home_screen.dart';