+++ lib/features/point_of_sale/presentation/widgets/aside_menu/menu.dart
@@ -20,6 +20,7 @@ class AsideMenu extends ConsumerWidget {
       ),
       child: Column(
         children: [
+          Gap(20),
           FlutterLogo(
             size: 40, 
             style: FlutterLogoStyle.markOnly,