+++ lib/features/point_of_sale/presentation/widgets/header_search_bar/header_search_bar_widget.dart
@@ -8,12 +8,17 @@ class HeaderSearchBar extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
 
-    return Row(
-      spacing: 15,
-      children: [
-        SearchBar(),
-        ProfileCard()
-      ],
+    return Padding(
+      padding: EdgeInsets.symmetric(
+        horizontal: 20
+      ),
+      child: Row(
+        spacing: 15,
+        children: [
+          SearchBar(),
+          ProfileCard()
+        ],
+      ),
     );
 
   }