+++ lib/features/point_of_sale/presentation/widgets/products/products_widget.dart
@@ -11,7 +11,9 @@ class ProductsWidget extends StatelessWidget {
 
     return Container(
       padding: EdgeInsets.only(
-        top: 20
+        top: 20,
+        left: 20,
+        right: 20
       ),
       child: const Column(
         spacing: 15,